+++ hain_tv/lib/player/exo_player_backend.dart
@@ -71,5 +71,8 @@ class ExoPlayerBackend implements VideoPlayerBackend {
   Stream<bool> get playingStream => _impl.playingStream;
 
   @override
+  Stream<void> get completedStream => _impl.completedStream;
+
+  @override
   Future<void> dispose() => _impl.dispose();
 }